split up python-twitch
