+++ twitch/keys.py
@@ -1,0 +1,47 @@
+# -*- encoding: utf-8 -*-
+"""
+    keys
+
+    string constants
+"""
+
+BITRATE = 'bitrate'
+CHANNEL = 'channel'
+CHANNELS = 'channels'
+CONNECT = 'connect'
+BACKGROUND = 'background'
+DISPLAY_NAME = 'display_name'
+FEATURED = 'featured'
+FOLLOWS = 'follows'
+GAME = 'game'
+GAMES = 'games'
+LOGO = 'logo'
+BOX = 'box'
+LARGE = 'large'
+NAME = 'name'
+NEEDED_INFO = 'needed_info'
+PLAY = 'play'
+PLAYPATH = 'playpath'
+QUALITY = 'quality'
+RTMP = 'rtmp'
+STREAMS = 'streams'
+REFERER = 'Referer'
+RTMP_URL = 'rtmpUrl'
+STATUS = 'status'
+STREAM = 'stream'
+SWF_URL = 'swfUrl'
+TEAMS = 'teams'
+TOKEN = 'token'
+TOP = 'top'
+TOTAL = '_total'
+USER_AGENT = 'User-Agent'
+USER_AGENT_STRING = ('Mozilla/5.0 (Windows NT 6.1; WOW64; rv:6.0) '
+                     'Gecko/20100101 Firefox/6.0')
+VIDEOS = 'videos'
+VIDEO_BANNER = 'video_banner'
+VIDEO_HEIGHT = 'video_height'
+VIEWERS = 'viewers'
+PREVIEW = 'preview'
+TITLE = 'title'
+QUALITY_LIST_STREAM = ['Source', 'High', 'Medium', 'Low', 'Mobile']
+QUALITY_LIST_VIDEO = ['live', '720p', '480p', '360p', '226p']
